Refactor: Template list page; HTTP API calls centralised
