+++ app/src/pages/favouriteMoviesPage.jsx
@@ -1,7 +1,19 @@
 import React from "react";
+import PageTemplate from "../components/templateMovieListPage";
 
-const FavouriteMoviesPage = () => {
-    return <h2>Favourite Movies</h2>
-}
+const FavouriteMoviesPage = (props) => {
+  const toDo = () => true;
+  // Get movies from local storage.
+  const movies = JSON.parse(localStorage.getItem("favourites")); 
 
-export default FavouriteMoviesPage
+  return (
+    <PageTemplate
+      title="Favourite Movies"
+      movies={movies}
+      selectFavourite={toDo}
+    />
+  );
+};
+
+export default FavouriteMoviesPage;
+
