+++ app/src/pages/homePage.jsx
@@ -1,44 +1,11 @@
 import React, { useState, useEffect } from "react";
-import Header from "../components/headerMovieList";
-import FilterCard from "../components/filterMoviesCard";
-import MovieList from "../components/movieList";
-import Grid from "@mui/material/Grid";
-import Fab from "@mui/material/Fab";
-import Drawer from "@mui/material/Drawer";
+import PageTemplate from '../components/templateMovieListPage'
+import { getMovies } from "../api/tmdb-api";
 
-
-const styles = {
-  root: {
-    padding: "20px",
-  },
-  fab: {
-    marginTop: 8,
-    position: "fixed",
-    top: 2,
-    right: 2,
-  },
-};
-
-const MovieListPage = (props) => {
+const HomePage = (props) => {
   const [movies, setMovies] = useState([]);
-  const [titleFilter, setTitleFilter] = useState("");
-  const [genreFilter, setGenreFilter] = useState("0");
-  const [drawerOpen, setDrawerOpen] = useState(false);
-
-  const genreId = Number(genreFilter);
-
-  let displayedMovies = movies
-    .filter((m) => {
-      return m.title.toLowerCase().search(titleFilter.toLowerCase()) !== -1;
-    })
-    .filter((m) => {
-      return genreId > 0 ? m.genre_ids.includes(genreId) : true;
-    });
-
-  const handleChange = (type, value) => {
-    if (type === "title") setTitleFilter(value);
-    else setGenreFilter(value);
-  };
+  const favourites = movies.filter(m => m.favourite)
+  localStorage.setItem('favourites', JSON.stringify(favourites))
 
   const addToFavourites = (movieId) => {
     const updatedMovies = movies.map((m) =>
@@ -48,51 +15,20 @@ const MovieListPage = (props) => {
   };
 
   useEffect(() => {
-    fetch(
-      `https://api.themoviedb.org/3/discover/movie?api_key=${import.meta.env.VITE_TMDB_KEY}&language=en-US&include_adult=false&page=1`
-    )
-      .then((res) => res.json())
-      .then((json) => {
-        // console.log(json);
-        return json.results;
-      })
-      .then((movies) => {
-        setMovies(movies);
-      });
+    getMovies().then(movies => {
+      setMovies(movies);
+    });
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
-    <>
-      <Grid container sx={styles.root}>
-        <Grid item xs={12}>
-          <Header title={"Home Page"} />
-        </Grid>
-        <Grid item container spacing={5}>
-          <MovieList movies={displayedMovies} selectFavourite={addToFavourites} />
-        </Grid>
-      </Grid>
-      <Fab
-          color="secondary"
-          variant="extended"
-          onClick={() => setDrawerOpen(true)}
-          sx={styles.fab}
-        >
-          Filter
-      </Fab>
-      <Drawer
-        anchor="left"
-        open={drawerOpen}
-        onClose={() => setDrawerOpen(false)}
-      >
-        <FilterCard
-          onUserInput={handleChange}
-          titleFilter={titleFilter}
-          genreFilter={genreFilter}
-        />
-      </Drawer>
-    </>
+    <PageTemplate
+      title='Discover Movies'
+      movies={movies}
+      selectFavourite={addToFavourites}
+    />
   );
 };
-export default MovieListPage;
+export default HomePage;
+
